refactor: 애플리케이션 DTO 금액 타입 Long → BigDecimal 변경
- SettlementItemCreateCommand: grossAmount 변경
- PartialSettlementAvailableItemResult: grossAmount, feeAmount, netAmount 변경
- PartialSettlementCreateResult: totalSalesAmount, feeAmount, finalSettlementAmount 변경
- PartialSettlementExecutionResult: totalSalesAmount, feeAmount, finalSettlementAmount 변경

+++ settlement/src/main/java/com/example/settlement/application/dto/PartialSettlementExecutionResult.java
@@ -2,6 +2,7 @@ package com.example.settlement.application.dto;
 
 import com.example.settlement.domain.enumtype.SettlementStatus;
 import com.example.settlement.domain.enumtype.SettlementType;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -13,8 +14,8 @@ public record PartialSettlementExecutionResult(
         SettlementType settlementType,
         SettlementStatus settlementStatus,
         int settlementItemCount,
-        Long totalSalesAmount,
-        Long feeAmount,
-        Long finalSettlementAmount
+        BigDecimal totalSalesAmount,
+        BigDecimal feeAmount,
+        BigDecimal finalSettlementAmount
 ) {
 }
